+++ index.html
@@ -228,6 +228,6 @@
     </div>
   </template>
 
-  <script defer src="js/main.js"></script>
+  <script defer type="module" src="js/main.js"></script>
 </body>
 </html>
